Renders a static threejs scene
* Scene => container for all objects & lights that we want to render
* Camera => defines what we'll see when we render the scene
* We use WebGLRenderer to utilise graphics card to render
* Adds Axes & plane for visual debugging
* Renderer does the math on how to render the scene based on camera's angle

We build a scene by creating objects & setting properties before adding the object to the scene in order for it to be rendered.

In this example, we rendered a simple sphere & cube using a `MeshBasicMaterial`, a simplest kind of material which does nothing fancy, we also render it as wireframes.

We also adjusted camera properties in a way so that we're viewing the objects from up above them.

Any angles that we set must be set in terms of radians (Math.PI = 180 degrees)