Init lights
We start off with ambient light. An ambient light is the kind of light that illuminates each & every pixel equally & it doesn't cast shadow. An ambient in itself is not enough for a scene & hence we also included a spotlight in this example so that we could see the sphere & cube.

A spotlight on the other hand, illuminates things in a cone shaped direction, more like a desk lamp. It does cast shadows.

Why do we need ambient light then? An usecase would probably be rendering a scene which is supposed to portray hot temperature or cold temperature, we could use a orangish-red colored ambient light for a sunny day & blue colored ambient light for a winter-y day.